+++ MicroMouse-2016.vcxproj
@@ -157,6 +157,7 @@
     <ClCompile Include="micromouse\MouseBot.cpp" />
     <ClCompile Include="micromouse\Node.cpp" />
     <ClCompile Include="micromouse\Path.cpp" />
+    <ClCompile Include="micromouse\PIDControl.cpp" />
     <ClCompile Include="micromouse\RobotIO.cpp" />
     <ClCompile Include="micromouse\IRSensor.cpp" />
     <ClCompile Include="micromouse\Vector.cpp" />
@@ -171,6 +172,7 @@
     <ClInclude Include="micromouse\MouseBot.h" />
     <ClInclude Include="micromouse\Node.h" />
     <ClInclude Include="micromouse\Path.h" />
+    <ClInclude Include="micromouse\PIDControl.h" />
     <ClInclude Include="micromouse\RobotIO.h" />
     <ClInclude Include="micromouse\IRSensor.h" />
     <ClInclude Include="micromouse\Vector.h" />
