+++ MicroMouse-2016.vcxproj.filters
@@ -51,6 +51,9 @@
     <ClCompile Include="micromouse\Motor.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="micromouse\PIDControl.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="micromouse\Controller.h">
@@ -87,6 +90,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="micromouse\Motor.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="micromouse\PIDControl.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
